Wire up logging.level and log_ignored_messages; fix IPv4 shift portability
logging.level was loaded but silently ignored — the DPP on_log threshold
was hardcoded to ll_warning. Now routes through a parseLogLevel() helper
so operators can tune to TRACE/DEBUG/INFO/WARNING/ERROR/CRITICAL. Default
changed from "INFO" to "WARNING" to preserve existing behaviour.

log_ignored_messages was parsed but never read. Now logs a debug line in
on_message_create for each filtered message (bot, webhook, wrong guild,
non-relay channel).

parseIpv4Address: shift operands cast to uint32_t before shifting so the
intermediate expression is well-defined on platforms where unsigned int is
narrower than 32 bits (currently UB if unsigned int < 32 bits).

+++ src/bridge_support.cpp
@@ -109,7 +109,10 @@ std::optional<std::uint32_t> parseIpv4Address(const std::string &value)
         }
     }
 
-    return (octets[0] << 24U) | (octets[1] << 16U) | (octets[2] << 8U) | octets[3];
+    return (static_cast<std::uint32_t>(octets[0]) << 24U) |
+           (static_cast<std::uint32_t>(octets[1]) << 16U) |
+           (static_cast<std::uint32_t>(octets[2]) <<  8U) |
+            static_cast<std::uint32_t>(octets[3]);
 }
 
 std::string normalizeAvatarProvider(std::string value)
